remove unnecessary FramedMmap
